requirements.txt
Pyrogram==2.0.32
TgCrypto
python-dotenv==0.20.0
PyYAML==6.0
